[P] Implemented -f/--fragmentation, -t/--top options; revised the -0/—freespace and -s/—summary options.
Added a -v/--verbose option to alter the verbosity level of the output of the -0/--freespace and -f/--fragmentation options. This still allows for the -v option by itself to be interpreted as "version", provided it's the only option specified. If there's more than one option specified, -v is interpreted to mean "verbose". (This is similar to how `codesign` uses the -v flag).

Added a -t/--top command-line option to specify the number of largest contiguous freespace segments to display (when used with the -0/--freespace option), the number of most fragmented files to list (when used with the -f/--fragmentation option), or the number of largest files to display (when used with the -s/--summary option) (similar to `hfsdebug`).

Implemented a -f/--fragmentation option to show fragmented files. The -f option by itself will simply show an overall summary of volume fragmentation. If you add the -t option, it will list the specified number of most-fragmented files. Finally, if you add the -v/--verbose option, it will list _all_ fragmented files, in a manner similar to `hfsdebug`.

Altered the -s/--summary option to use the -t/--top option to allow a user-specified number of largest files to display in the summary.

NOTE: Slightly altered the VolumeSummary procedure, as merely because a file is invisible or is an alias, doesn't mean it can't be fragmented.

Changed the Volume Summary printing procedure to print the full path to the file rather than just the file name.

Altered the -0/--freespace option to use the -t/--top and -v/--verbose options. You can use the -t option to quickly see the largest segments of freespace available, and adding the -v option will list all freespace segments, like `hfsdebug` does.

Updated the man page and built-in tool help as best as I could.

Added a Debug and a Release Xcode scheme.